% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scFeatureView.R
\name{mod_scFeatureView_ui}
\alias{mod_scFeatureView_ui}
\title{scFeatureView UI Function}
\usage{
mod_scFeatureView_ui(id, title)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
